requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.31.0
duckduckgo-search==6.1.0
google-generativeai==0.8.4